requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
torchvision==0.21.0+cu124
numpy==2.2.6
pandas==2.3.1
Pillow==11.0.0
scikit-learn==1.7.2
tqdm==4.67.1
regex==2026.1.15
